Seen_Pipeline/gdrnpp_bop2022/requirements/requirements.txt
# PyTorch and related packages with CUDA 11.6 support
torch==1.13.1+cu116
torchvision==0.14.1+cu116
torchaudio==0.13.1+cu116
-f https://download.pytorch.org/whl/torch_stable.html

# Core dependencies
cython
plyfile
ujson
pycocotools
cffi
ninja
black
docformatter
setproctitle
meshplex
OpenEXR
vispy>=0.6.4
yacs>=0.1.8
tabulate
pytest-runner
pytest
ipdb
tqdm
numba

# MMCV-full with compatible CUDA
mmcv-full
-f https://download.openmmlab.com/mmcv/dist/cu116/torch1.13/index.html

# Additional dependencies
pyassimp==4.1.3  # Version 4.1.4 causes SegmentFault in egl_renderer
pypng
imgaug>=0.4.0
albumentations
transforms3d
open3d
fvcore
tensorboardX
einops
pytorch3d
timm==0.6.7
glfw
imageio
imageio-ffmpeg
PyOpenGL>=3.1.5
PyOpenGL_accelerate
chardet

# PyTorch-related utilities
thop  # https://github.com/Lyken17/pytorch-OpCounter
loguru
pytorch-lightning
fairscale
deepspeed

# ONNX and related packages
pyro-ppl==1.4.0  # For epro-pnp <1.8.1
onnx==1.8.1
onnxruntime==1.8.0
onnx-simplifier==0.3.5
